moves formatDate to own file, refactors searched films, renders in table, applies styling
