--- src/components/movies/Movies.css
@@ -1,4 +1,0 @@
-.container{
-    padding-top: 40px;
-    padding-bottom: 10px;
-}
