+++ src/components/savedFilm/SavedFilm.js
@@ -26,6 +26,7 @@ import EditIcon from "@mui/icons-material/Edit";
 import FavoriteIcon from "@mui/icons-material/Favorite";
 import LocalMoviesIcon from "@mui/icons-material/LocalMovies";
 import { StyledTableCell, StyledTableRow, style } from "./styles";
+import { formatDate } from "../../formatDate";
 
 export default function SavedFilm() {
   const {
@@ -42,36 +43,36 @@ export default function SavedFilm() {
   const [openEdit, setOpenEdit] = useState(false);
   const [selectedMovieEdit, setSelectedMovieEdit] = useState(null);
 
-  const formatDate = (date, format) => {
-    const dateObj = new Date(date);
+  // const formatDate = (date, format) => {
+  //   const dateObj = new Date(date);
 
-    if (isNaN(dateObj)) {
-      return "Invalid date";
-    }
+  //   if (isNaN(dateObj)) {
+  //     return "Invalid date";
+  //   }
 
-    switch (format) {
-      case "day":
-        return dateObj.toLocaleDateString(undefined, {
-          day: "numeric",
-        });
-      case "year":
-        return dateObj.toLocaleDateString(undefined, {
-          year: "numeric",
-        });
-      case "monthAndYear":
-        return dateObj.toLocaleDateString(undefined, {
-          month: "short",
-          year: "numeric",
-        });
-      case "default":
-        const month = `${dateObj.getMonth() + 1}`.padStart(2, "0");
-        const day = `${dateObj.getDate()}`.padStart(2, "0");
-        const year = dateObj.getFullYear();
-        return [year, month, day].join("-");
-      default:
-        throw new Error(`Invalid date format: ${format}`);
-    }
-  };
+  //   switch (format) {
+  //     case "day":
+  //       return dateObj.toLocaleDateString(undefined, {
+  //         day: "numeric",
+  //       });
+  //     case "year":
+  //       return dateObj.toLocaleDateString(undefined, {
+  //         year: "numeric",
+  //       });
+  //     case "monthAndYear":
+  //       return dateObj.toLocaleDateString(undefined, {
+  //         month: "short",
+  //         year: "numeric",
+  //       });
+  //     case "default":
+  //       const month = `${dateObj.getMonth() + 1}`.padStart(2, "0");
+  //       const day = `${dateObj.getDate()}`.padStart(2, "0");
+  //       const year = dateObj.getFullYear();
+  //       return [year, month, day].join("-");
+  //     default:
+  //       throw new Error(`Invalid date format: ${format}`);
+  //   }
+  // };
 
   const handleOpenEdit = (id) => {
     const findId = savedMovies?.find((savedFilm) => savedFilm._id === id);
